# exercise 8/part 1/calculator-api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Calculator API",
    "version": "1.0.0",
    "description": "API documentation for Calculator endpoints"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/calculator/add": {
      "get": {
        "summary": "Add two numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "required": true,
            "schema": { "type": "number" }
          },
          {
            "name": "num2",
            "in": "query",
            "required": true,
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful addition",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": { "type": "number" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/calculator/subtract": {
      "get": {
        "summary": "Subtract two numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "required": true,
            "schema": { "type": "number" }
          },
          {
            "name": "num2",
            "in": "query",
            "required": true,
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful subtraction",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": { "type": "number" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/calculator/multiply": {
      "get": {
        "summary": "Multiply two numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "required": true,
            "schema": { "type": "number" }
          },
          {
            "name": "num2",
            "in": "query",
            "required": true,
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful multiplication",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": { "type": "number" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/calculator/divide": {
      "get": {
        "summary": "Divide two numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "required": true,
            "schema": { "type": "number" }
          },
          {
            "name": "num2",
            "in": "query",
            "required": true,
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful division",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": { "type": "number" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Division by zero error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}